Rename shared JITX package to shared-components
