+++ jitx-py/alchitry-au1-level-shifter/src/main.py
@@ -17,9 +17,9 @@ from jitx.shapes.composites import rectangle
 from jitx.shapes.primitive import Circle, Text
 from jitx.substrate import Substrate
 from jitx.via import Via, ViaType
-from retrobus_jitx_parts.ffc import HDGC60PinFfc
-from retrobus_jitx_parts.saleae import SaleaeProbeHeader2x4
-from retrobus_jitx_parts.testpads import GndTestpads
+from shared_components.ffc import HDGC60PinFfc
+from shared_components.saleae import SaleaeProbeHeader2x4
+from shared_components.testpads import GndTestpads
 
 from src.components import (
     AlchitryA,
