+++ jitx-py/pin-tester/src/main.py
@@ -15,8 +15,8 @@ from jitx.shapes.composites import rectangle
 from jitx.shapes.primitive import Circle, Text
 from jitx.substrate import Substrate
 from jitx.via import Via, ViaType
-from retrobus_jitx_parts.ffc import HDGC60PinFfc
-from retrobus_jitx_parts.testpads import GndTestpads
+from shared_components.ffc import HDGC60PinFfc
+from shared_components.testpads import GndTestpads
 
 from src.components import (
     PinHeader2x4,
