--- jitx-py/retrobus-jitx-parts/src/retrobus_jitx_parts/__init__.py
@@ -1,11 +1,0 @@
-from retrobus_jitx_parts.ffc import HDGC60PinFfc
-from retrobus_jitx_parts.saleae import SaleaeProbeHeader2x4, SignalGroundHeader2x4
-from retrobus_jitx_parts.testpads import GndTestpads, SignalTestPad
-
-__all__ = [
-    "GndTestpads",
-    "HDGC60PinFfc",
-    "SaleaeProbeHeader2x4",
-    "SignalGroundHeader2x4",
-    "SignalTestPad",
-]
