+++ jitx-py/saleae-dslab-adapter/src/main.py
@@ -14,7 +14,7 @@ from jitx.sample import SampleFabConstraints, SampleStackup
 from jitx.shapes.composites import rectangle
 from jitx.shapes.primitive import Text
 from jitx.substrate import Substrate
-from retrobus_jitx_parts.saleae import SaleaeProbeHeader2x4
+from shared_components.saleae import SaleaeProbeHeader2x4
 
 from src.components import DSLabFemaleHeader2x4
