+++ jitx-py/shared-components/README.md
@@ -1,3 +1,3 @@
-# retrobus-jitx-parts
+# shared-components
 
 Shared reusable JITX Python parts for the `jitx-py` board projects.
