+++ jitx-py/shared-components/src/shared_components/__init__.py
@@ -1,0 +1,11 @@
+from shared_components.ffc import HDGC60PinFfc
+from shared_components.saleae import SaleaeProbeHeader2x4, SignalGroundHeader2x4
+from shared_components.testpads import GndTestpads, SignalTestPad
+
+__all__ = [
+    "GndTestpads",
+    "HDGC60PinFfc",
+    "SaleaeProbeHeader2x4",
+    "SignalGroundHeader2x4",
+    "SignalTestPad",
+]
